+++ db/migrations/20160502172140-create-table-production_unit.ts
@@ -63,7 +63,7 @@ module.exports = {
             allowNull: true
         },
         day_rate: {
-            type: Sequelize.FLOAT,
+            type: Sequelize.DOUBLE,
 	          allowNull: true
         },
 				photo: {
